Add OAuthInterceptor to the Retorfit client
OAuthInterceptor is an Okhttp3 interceptor which appends the Battle.net
OAuth access token to each Battle.net API request, if one is not already
present. If the access token is not known, it will retrieve it from the
Battle.net OAuth API and store it on the instance for future use.