Add interfaces and base classes for WoW Indexes
Index-type API responses in the WoW API contain "minimal" versions of
full API items. For example, the Achievement Index API will return a
list of all achievements, but only the "id" and "name" field.

Each one of these index-type items also contains a "key" field which
holds a permalink to the API endpoing that will return the full version.
Calling GET on this URL will return a full version of that item with all
of the fields.

This is represented in an interface by `Keyed<T>`. The biggest thing
here is that this interface contains a `Optional<T> getFullItem()`
method. Implementing subclasses should define the type `T` of the full
item that will be returned by calling that URL in the key.

The base implementation of this is contained in `AbstractWoWIndexItem`
which calls the API necessary and returns the built POJO model.