Switch API Responses to the Result library
Result is a Java implementation of the Rust language's Result enum type,
and provides a functional style of error handling. Provides the ability
to return an "error type" in the case that an API call fails, which
makes it easier to surface this error to the caller rather than long
series of exception handling or through logging adapters.