Adds support for multiple regions and locales
Adds support for US, EU, KR, and TW Battle.net API regions, and all of
their associated locales. A full list of these can be found here:
https://develop.battle.net/documentation/guides/regionality-partitions-and-localization

Additionally, adds support for WoW API namespaces. API Reference:
https://develop.battle.net/documentation/guides/game-data-apis-wow-namespaces

The region used in the base API urls is done at client construction with
the new builder methods. Since the locale and namespace are part of the
query parameters, it's difficult to define these statically with
Retrofit annotations.

Because of this, it is done via an okhttp3 interceptor. Rather than
adding multiple interceptors, this commit adds a master
`BattleNetInterceptor` which handles OAuth, Namespace, and Locale
additions to the request URL.

This commit should complete the majority of the "core" of the API
library. Short of a few *strange* models used by some of the APIs, the
next steps will be to add World of Warcraft Game Data and Profile APIs
which will be able to retrieve actual data.