Add an interface for Cache implementations
The cache can be used to store fetched API data for re-use
rather than needing to call the API on every access.

The cache can be customized to use different TTL/expires durations for
individual BattleNetEntity types or, in the case of WoW APIs, the
namespaces. For example, static namespace can be cached for longer than
dynamic or profile API entities.